+++ crates/trie/db/Cargo.toml
@@ -18,6 +18,7 @@ reth-execution-errors.workspace = true
 reth-db.workspace = true
 reth-db-api.workspace = true
 reth-stages-types.workspace = true
+reth-storage-errors.workspace = true
 reth-trie-common.workspace = true
 reth-trie.workspace = true
