+++ crates/trie/trie/src/state.rs
@@ -4,14 +4,7 @@ use crate::{
 };
 use itertools::Itertools;
 use rayon::prelude::{IntoParallelIterator, ParallelIterator};
-use reth_db::tables;
-use reth_db_api::{
-    cursor::DbCursorRO,
-    models::{AccountBeforeTx, BlockNumberAddress},
-    transaction::DbTx,
-};
-use reth_primitives::{keccak256, Account, Address, BlockNumber, B256, U256};
-use reth_storage_errors::db::DatabaseError;
+use reth_primitives::{keccak256, Account, Address, B256, U256};
 use revm::db::BundleAccount;
 use std::collections::{hash_map, HashMap, HashSet};
 
@@ -53,52 +46,6 @@ impl HashedPostState {
             storages.insert(address, storage);
         }
         Self { accounts, storages }
-    }
-
-    /// Initializes [`HashedPostState`] from reverts. Iterates over state reverts from the specified
-    /// block up to the current tip and aggregates them into hashed state in reverse.
-    pub fn from_reverts<TX: DbTx>(tx: &TX, from: BlockNumber) -> Result<Self, DatabaseError> {
-        // Iterate over account changesets and record value before first occurring account change.
-        let mut accounts = HashMap::<Address, Option<Account>>::default();
-        let mut account_changesets_cursor = tx.cursor_read::<tables::AccountChangeSets>()?;
-        for entry in account_changesets_cursor.walk_range(from..)? {
-            let (_, AccountBeforeTx { address, info }) = entry?;
-            if let hash_map::Entry::Vacant(entry) = accounts.entry(address) {
-                entry.insert(info);
-            }
-        }
-
-        // Iterate over storage changesets and record value before first occurring storage change.
-        let mut storages = HashMap::<Address, HashMap<B256, U256>>::default();
-        let mut storage_changesets_cursor = tx.cursor_read::<tables::StorageChangeSets>()?;
-        for entry in
-            storage_changesets_cursor.walk_range(BlockNumberAddress((from, Address::ZERO))..)?
-        {
-            let (BlockNumberAddress((_, address)), storage) = entry?;
-            let account_storage = storages.entry(address).or_default();
-            if let hash_map::Entry::Vacant(entry) = account_storage.entry(storage.key) {
-                entry.insert(storage.value);
-            }
-        }
-
-        let hashed_accounts = HashMap::from_iter(
-            accounts.into_iter().map(|(address, info)| (keccak256(address), info)),
-        );
-
-        let hashed_storages = HashMap::from_iter(storages.into_iter().map(|(address, storage)| {
-            (
-                keccak256(address),
-                HashedStorage::from_iter(
-                    // The `wiped` flag indicates only whether previous storage entries
-                    // should be looked up in db or not. For reverts it's a noop since all
-                    // wiped changes had been written as storage reverts.
-                    false,
-                    storage.into_iter().map(|(slot, value)| (keccak256(slot), value)),
-                ),
-            )
-        }));
-
-        Ok(Self { accounts: hashed_accounts, storages: hashed_storages })
     }
 
     /// Set account entries on hashed state.
